fix client feature flags, move error helper functions to MentatError
