+++ mentat/Cargo.toml
@@ -6,7 +6,7 @@ rust-version = "1.59.0"
 
 [features]
 _network = ["indexmap", "serde", "serde_json", "reqwest"]
-client = ["_network"]
+client = ["_network", "anyhow", "tokio"]
 default = ["server"]
 server = [
     "_network",
@@ -29,6 +29,7 @@ sdk = ["keys"]
 keys = ["ed25519-dalek", "mina-hasher", "mina-signer", "p256", "p256/ecdsa", "rand", "secp256k1", "secp256k1/bitcoin_hashes", "secp256k1/recovery", "secp256k1/rand-std", "snarkvm-dpc", "snarkvm-algorithms", "snarkvm-utilities"]
 
 [dependencies]
+anyhow = { version = "1.0.56", optional = true }
 axum = { version = "0.4.8", optional = true, default-features = false, features = ["http1", "json"] }
 color-backtrace = { version = "0.5.1", optional = true, default-features = false }
 ed25519-dalek = { version = "1.0", optional = true, default-features = false }
