+++ mentat/src/api/data.rs
@@ -1,4 +1,5 @@
 use super::*;
+use crate::errors::MentatError;
 
 #[axum::async_trait]
 pub trait DataApi: Default {
@@ -10,7 +11,7 @@ pub trait DataApi: Default {
         _data: MetadataRequest,
         _rpc_caller: RpcCaller,
     ) -> MentatResponse<NetworkListResponse> {
-        ApiError::not_implemented()
+        MentatError::not_implemented()
     }
 
     /// This endpoint returns the version information and allowed
@@ -24,7 +25,7 @@ pub trait DataApi: Default {
         _data: NetworkRequest,
         _rpc_caller: RpcCaller,
     ) -> MentatResponse<NetworkOptionsResponse> {
-        ApiError::not_implemented()
+        MentatError::not_implemented()
     }
 
     /// This endpoint returns the current status of the network requested. Any
@@ -35,7 +36,7 @@ pub trait DataApi: Default {
         _data: NetworkRequest,
         _rpc_caller: RpcCaller,
     ) -> MentatResponse<NetworkStatusResponse> {
-        ApiError::not_implemented()
+        MentatError::not_implemented()
     }
 
     /// Get an array of all AccountBalances for an AccountIdentifier and the
@@ -59,7 +60,7 @@ pub trait DataApi: Default {
         _data: AccountBalanceRequest,
         _rpc_caller: RpcCaller,
     ) -> MentatResponse<AccountBalanceResponse> {
-        ApiError::not_implemented()
+        MentatError::not_implemented()
     }
 
     /// Get an array of all unspent coins for an AccountIdentifier and the
@@ -84,7 +85,7 @@ pub trait DataApi: Default {
         _data: AccountCoinsRequest,
         _rpc_caller: RpcCaller,
     ) -> MentatResponse<AccountCoinsResponse> {
-        ApiError::not_implemented()
+        MentatError::not_implemented()
     }
 
     /// Get a block by its Block Identifier. If transactions are returned in the
@@ -103,7 +104,7 @@ pub trait DataApi: Default {
         _data: BlockRequest,
         _rpc_caller: RpcCaller,
     ) -> MentatResponse<BlockResponse> {
-        ApiError::not_implemented()
+        MentatError::not_implemented()
     }
 
     /// Get a transaction in a block by its Transaction Identifier. This
@@ -130,7 +131,7 @@ pub trait DataApi: Default {
         _data: BlockTransactionRequest,
         _rpc_caller: RpcCaller,
     ) -> MentatResponse<BlockTransactionResponse> {
-        ApiError::not_implemented()
+        MentatError::not_implemented()
     }
 
     /// Get all Transaction Identifiers in the mempool
@@ -140,7 +141,7 @@ pub trait DataApi: Default {
         _data: NetworkRequest,
         _rpc_caller: RpcCaller,
     ) -> MentatResponse<MempoolResponse> {
-        ApiError::not_implemented()
+        MentatError::not_implemented()
     }
 
     /// Get a transaction in the mempool by its Transaction Identifier. This is
@@ -158,7 +159,7 @@ pub trait DataApi: Default {
         _data: MempoolTransactionRequest,
         _rpc_caller: RpcCaller,
     ) -> MentatResponse<MempoolTransactionResponse> {
-        ApiError::not_implemented()
+        MentatError::not_implemented()
     }
 }
 
@@ -192,7 +193,7 @@ pub trait CallerDataApi: Clone + DataApi {
         rpc_caller: RpcCaller,
     ) -> MentatResponse<NetworkStatusResponse> {
         if mode.is_offline() {
-            ApiError::wrong_network(&data)
+            MentatError::wrong_network(&data)
         } else {
             self.network_status(caller, data, rpc_caller).await
         }
@@ -206,7 +207,7 @@ pub trait CallerDataApi: Clone + DataApi {
         rpc_caller: RpcCaller,
     ) -> MentatResponse<AccountBalanceResponse> {
         if mode.is_offline() {
-            ApiError::wrong_network(&data)
+            MentatError::wrong_network(&data)
         } else {
             self.account_balance(caller, data, rpc_caller).await
         }
@@ -220,7 +221,7 @@ pub trait CallerDataApi: Clone + DataApi {
         rpc_caller: RpcCaller,
     ) -> MentatResponse<AccountCoinsResponse> {
         if mode.is_offline() {
-            ApiError::wrong_network(&data)
+            MentatError::wrong_network(&data)
         } else {
             self.account_coins(caller, data, rpc_caller).await
         }
@@ -234,7 +235,7 @@ pub trait CallerDataApi: Clone + DataApi {
         rpc_caller: RpcCaller,
     ) -> MentatResponse<BlockResponse> {
         if mode.is_offline() {
-            ApiError::wrong_network(&data)
+            MentatError::wrong_network(&data)
         } else {
             self.block(caller, data, rpc_caller).await
         }
@@ -248,7 +249,7 @@ pub trait CallerDataApi: Clone + DataApi {
         rpc_caller: RpcCaller,
     ) -> MentatResponse<BlockTransactionResponse> {
         if mode.is_offline() {
-            ApiError::wrong_network(&data)
+            MentatError::wrong_network(&data)
         } else {
             self.block_transaction(caller, data, rpc_caller).await
         }
@@ -262,7 +263,7 @@ pub trait CallerDataApi: Clone + DataApi {
         rpc_caller: RpcCaller,
     ) -> MentatResponse<MempoolResponse> {
         if mode.is_offline() {
-            ApiError::wrong_network(&data)
+            MentatError::wrong_network(&data)
         } else {
             self.mempool(caller, data, rpc_caller).await
         }
@@ -276,7 +277,7 @@ pub trait CallerDataApi: Clone + DataApi {
         rpc_caller: RpcCaller,
     ) -> MentatResponse<MempoolTransactionResponse> {
         if mode.is_offline() {
-            ApiError::wrong_network(&data)
+            MentatError::wrong_network(&data)
         } else {
             self.mempool_transaction(caller, data, rpc_caller).await
         }
