+++ rosetta-bitcoin/src/data_api.rs
@@ -128,7 +128,7 @@ impl DataApi for BitcoinDataApi {
                 transaction: tx.into_transaction(i, &rpc_caller).await?,
             }))
         } else {
-            ApiError::unable_to_find_transaction(&data.transaction_identifier.hash)
+            MentatError::unable_to_find_transaction(&data.transaction_identifier.hash)
         }
     }
 
@@ -175,7 +175,7 @@ impl DataApi for BitcoinDataApi {
                 metadata: IndexMap::new(),
             }))
         } else {
-            ApiError::unable_to_find_transaction(&data.transaction_identifier.hash)
+            MentatError::unable_to_find_transaction(&data.transaction_identifier.hash)
         }
     }
 }
